requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
aiohttp==3.9.5
beautifulsoup4==4.12.3
